refactor: extract Violation type to rules/types.ts
