+++ src/rules/cohesion.ts
@@ -1,5 +1,5 @@
 import { ArchitectureModel, Boundary } from "../model";
-import { Violation } from "./acl";
+import { Violation } from "./types";
 
 export interface CohesionOptions {
   externalType?: string;
