+++ src/rules/crud.ts
@@ -1,5 +1,5 @@
 import { Container } from "../model";
-import { Violation } from "./acl";
+import { Violation } from "./types";
 
 export interface CrudOptions {
   repoTags?: string[];
