+++ src/rules/fixDbPerService.ts
@@ -1,5 +1,5 @@
 import type { ArchitectureModel } from "../model";
-import type { Violation } from "./acl";
+import type { Violation } from "./types";
 import type { DbPerServiceOptions } from "./dbPerService";
 import type { FixResult, SourceSyntax } from "./fix";
