+++ src/rules/types.ts
@@ -1,0 +1,4 @@
+export interface Violation {
+  container: string;
+  message: string;
+}
